+++ NEWS.md
@@ -43,6 +43,7 @@ changes.
 
 * `Array` is no more and its use has been replaced by `std::array`.
 * `Progress` is no longer a class template.
+* The Integrate module is removed.
 
 #Changes since v1 .1.1